fix: address builder snapshot review findings
